Fix hymie bugs
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@2415 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 70d5d0017069c72d549392ea1fe5220767b2a5bb
Former-commit-id: d0dbc48cc519d6631ef1d335f61e5b0bc998b89c

+++ src/drivers/hymie/driver.cpp
@@ -5904,7 +5904,7 @@ v2d Driver::getTargetPoint()
   }
 
   if ((avoiding & AVOID_FRONT) && !(avoiding & AVOID_SIDE))
-   offset = (float)myoffset = (float)MAX(-seg->width * 0.5f + 0.2f, MIN(seg->width*0.5f-0.2f, offset));
+   offset = myoffset = (float)MAX(-seg->width * 0.5f + 0.2f, MIN(seg->width*0.5f-0.2f, offset));
  }
 
  oldlookahead = (float)(lookahead * factor);
@@ -5932,11 +5932,11 @@ v2d Driver::getTargetPoint()
       ((track->pits.side == TR_LFT && nextoffset > offset) ||
        (track->pits.side == TR_RGT && nextoffset < offset)))
   {
-   offset = (float)myoffset = (float)nextoffset;
+   offset = myoffset = (float)nextoffset;
   }
   else
   {
-   nextoffset = (float)myoffset = (float)offset;
+   nextoffset = myoffset = (float)offset;
   }
  }
